added global top bar
